Separated mcc and ccm modules
