+++ modules/ccm/perf/perf_main.cpp
@@ -1,0 +1,3 @@
+#include "perf_precomp.hpp"
+
+CV_PERF_TEST_MAIN(ccm)
