+++ modules/ccm/perf/perf_precomp.hpp
@@ -2,11 +2,11 @@
 #define __OPENCV_PERF_PRECOMP_HPP__
 
 #include "opencv2/ts.hpp"
-#include "opencv2/mcc.hpp"
+#include "opencv2/ccm.hpp"
 
 namespace opencv_test
 {
-using namespace cv::mcc;
+using namespace cv::ccm;
 using namespace perf;
 }
