+++ modules/ccm/src/color.cpp
