+++ modules/ccm/src/precomp.hpp
@@ -26,29 +26,19 @@
  * SOFTWARE.
  */
 
-#ifndef __OPENCV_MCC_HPP__
-#define __OPENCV_MCC_HPP__
+#ifndef _CCM_PRECOMP_HPP
+#define _CCM_PRECOMP_HPP
 
+#include <limits>
 
-#include "mcc/checker_detector.hpp"
-#include "mcc/checker_model.hpp"
+#include <opencv2/core.hpp>
+#include <opencv2/imgproc.hpp>
+#include <opencv2/3d.hpp>
+#include <opencv2/dnn.hpp>
 
-/** @defgroup mcc Macbeth Chart module
-@{
+#include <vector>
+#include <string>
 
-Introduction
-------------
+#include "opencv2/ccm.hpp"
 
-ColorCharts are a tool for calibrating the color profile of camera, which not
-only depends on the intrinsic and extrinsic parameters of camera but also on the
-lighting conditions. This is done by taking the image of a chart, such that the
-value of its colors present in it known, in the image the color values changes
-depeding on many variables, this gives us the colors initially present and the
-colors that are present in the image, based on this information we can apply any
-suitable algorithm to find the actual color of all the objects present in the
-image.
-
-@}
-*/
-
-#endif
+#endif //_MCC_PRECOMP_HPP
