+++ modules/ccm/src/utils.cpp
